+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - refactor: error log to throw error
 - refactor: util method
 - refactor: move API `process` to `workflow`
+- feat: `getHelper` API
 
 ## 3.2.1